Refactoring Code ^^
