+++ client/src/components/UserDashboard/Toolbar.tsx
@@ -1,0 +1,38 @@
+import React from 'react';
+import AppBar from '@mui/material/AppBar';
+import CssBaseline from '@mui/material/CssBaseline';
+import IconButton from '@mui/material/IconButton';
+import MenuIcon from '@mui/icons-material/Menu';
+import Toolbar from '@mui/material/Toolbar';
+import Typography from '@mui/material/Typography';
+import { drawerWidth } from 'utils/settings';
+
+const ToolbarComponent: React.FC<{
+  handleDrawerToggle: () => void;
+}> = ({ handleDrawerToggle }) => (
+  <>
+    <CssBaseline />
+    <AppBar
+      position="fixed"
+      sx={{
+        width: { sm: `calc(100% - ${drawerWidth}px)` },
+        ml: { sm: `${drawerWidth}px` },
+      }}>
+      <Toolbar>
+        <IconButton
+          color="inherit"
+          aria-label="open drawer"
+          edge="start"
+          onClick={handleDrawerToggle}
+          sx={{ mr: 2, display: { sm: 'none' } }}>
+          <MenuIcon />
+        </IconButton>
+        <Typography variant="h6" noWrap component="div">
+          Responsive drawer
+        </Typography>
+      </Toolbar>
+    </AppBar>
+  </>
+);
+
+export default ToolbarComponent;
